refactor(ingest): consolidate default_emit_mode into DatahubClientConfig
Move the per-instance emit-mode default off the DataHubGraph constructor
sideband and onto DatahubClientConfig, so it's mapped like every other
emitter parameter and can be set declaratively in one place.

- Add default_emit_mode field to DatahubClientConfig; DataHubGraph reads
  it from config instead of taking a separate constructor argument
- Extract EmitMode into its own leaf module (datahub/emitter/emit_mode.py)
  so graph.config can reference it for the field without a circular import;
  rest_emitter re-exports it for backwards compatibility
- Dagster sensor forwards its emit_mode onto the client config
